itemselectionmodel: Move SymmetricSelectionModel to util
