+++ Orange/widgets/utils/itemselectionmodel.py
@@ -167,3 +167,41 @@ def ranges(indices):
         _, start = range_ind[0]
         _, end = range_ind[-1]
         yield start, end + 1
+
+
+class SymmetricSelectionModel(QItemSelectionModel):
+    def select(self, selection, flags):
+        if isinstance(selection, QModelIndex):
+            selection = QItemSelection(selection, selection)
+
+        model = self.model()
+        indexes = selection.indexes()
+        sel_inds = {ind.row() for ind in indexes} | \
+                   {ind.column() for ind in indexes}
+        if flags == QItemSelectionModel.ClearAndSelect:
+            selected = set()
+        else:
+            selected = {ind.row() for ind in self.selectedIndexes()}
+        if flags & QItemSelectionModel.Select:
+            selected |= sel_inds
+        elif flags & QItemSelectionModel.Deselect:
+            selected -= sel_inds
+        new_selection = QItemSelection()
+        regions = list(ranges(sorted(selected)))
+        for r_start, r_end in regions:
+            for c_start, c_end in regions:
+                top_left = model.index(r_start, c_start)
+                bottom_right = model.index(r_end - 1, c_end - 1)
+                new_selection.select(top_left, bottom_right)
+        QItemSelectionModel.select(self, new_selection,
+                                   QItemSelectionModel.ClearAndSelect)
+
+    def selected_items(self):
+        return list({ind.row() for ind in self.selectedIndexes()})
+
+    def set_selected_items(self, inds):
+        index = self.model().index
+        selection = QItemSelection()
+        for i in inds:
+            selection.select(index(i, i), index(i, i))
+        self.select(selection, QItemSelectionModel.ClearAndSelect)
